% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{R84KarBargh1}
\alias{R84KarBargh1}
\title{Table R84KarBargh1 \cr
Worksheet related to miscellaneous incomes of rural household members in 1384 \cr
كاربرگ مربوط به درامدهاي متفرقه اعضاء خانوار روستایی سال 1384}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف فرد}
\item Description: Person row number
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{شماره ردیف فعالیت}
\item Description: Activity row number
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{درآمد}
\item Description: Income
}
}
}
\usage{
R84KarBargh1
}
\description{
{R for Rural} \cr
{84 for 1384}
}
\keyword{datasets}
