% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{U84P3S13}
\alias{U84P3S13}
\title{Table U84P3S13 \cr
Supply and sale of durable goods & other yearly expenditure of urban household in 1384 \cr
تهيه و فروش كالاهای بادوام منزل شهری سال 1384}
\format{
A data frame with 6 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{تعداد افراد بیمه شده}
\item Description: Number of insured persons
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{خرید یا هزینه}
\item Description: Purchase or expenditure
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{فروش دست دوم}
\item Description: Second hand sale
}
}
}
\usage{
U84P3S13
}
\description{
{U for Urban} \cr
{84 for 1384} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
